Detailing the exact variants in the menu
Greetings. I just wanted to point out that the Spanish variant that was done on Transifex was Spanish (Spain). Since there are many other variants, I think it would be best to mark it properly in order to avoid any future translators that could start mixing different regions in the same language.

Thus, if someone wants to make a Spanish from another region, they could simply call for a new language (like that Chilean version that seems empty on Transifex) and do their own version without any mixes and mismatches.