Type Debugger: fallback for unknown types
Often the Typed Debugger seems to not be able to do anything about
custom enum types and such. In these cases, instead of hiding the value
all together, it's a lot nicer to allow editing the underlying bytes
anyway, at least assuming they fall into one of the blessed sizes.

Unsigned version of all the values was picked as it's the more "raw"
input. Especially if a value is actually a pointer or an enum, it's
usually not actually signed.

Incidentally, this makes types like Ghidra's `byte`, `undefined`,
`undefined4` etc just work out of the box without having to pre-process
them in the export script.