Workarounding gdb failure on connect.
If gdb connects while the emulator is running and printing lots of logs, the gdb client might get confused about getting messages when it's not supposed to. We're trying to avoid this situation by preventing messages to be sent until gdb may potentially be in a state where it's ready to accept them.